Update: breakpoint context update
